+++ src/timing.cpp
@@ -1,18 +1,18 @@
 #include <timing.h>
 
-static unsigned int ticks_ = 0;
+static uint64_t ticks_ = 0;
 
-unsigned int Timing::get_current_tick() {
+uint64_t Timing::get_current_tick() {
     return ticks_;
 }
 
 /// @brief 步进一周期
-void Timing::step_tick(unsigned int ticks) {
+void Timing::step_tick(uint64_t ticks) {
     ticks_ += ticks;
 }
 
 /// @brief 步进一CPU周期
-void Timing::step_cpu_tick(unsigned int ticks) {
+void Timing::step_cpu_tick(uint64_t ticks) {
     ticks_ += 3 * ticks;
 }
 
@@ -20,6 +20,6 @@ void Timing::reset_tick() {
     ticks_ = 0;
 }
 
-void Timing::set_tick(unsigned int ticks) {
+void Timing::set_tick(uint64_t ticks) {
     ticks_ = ticks;
 }
